% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Numeric integration}
\usage{
integrateIt(x, y, a, b, rule)
}
\arguments{
\item{x}{Vector containing x values}

\item{y}{Vector containing y values}

\item{rule}{specified rule (Trapezoid or Simpson)}
}
\value{
An object of class 'Simpson' or 'Trapezoid'

\itemize{
\item \code{x} Vector containing x values 
\item \code{y} Vector containing x values
\item \code{a} Lower limit of integration
\item \code{b} Upper limit of integration 
\item \code{rule} Specified rule (Trapezoid or Simpson).
}
}
\description{
Function creates an object of class 'Simpson' or 'Trapezoid' and estimatest numeric integration
}
\examples{

a=2
b=8
x <- c(1,5,6,7,2,3,9,4,8)
y=x^2
integrateIt(x=x, y=y, a=a, b=b, rule="Trapezoid")
integrateIt(x=x, y=y, a=a, b=b, rule="Simpson")

}
\author{
Mauricio Vela
}
\seealso{
\code{\link{Simpson}} \code{\link{Trapezoid}}
}

