% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateItpackage-package.R
\docType{package}
\name{integrateItPack}
\alias{integrateItPack}
\alias{integrateItPack-package}
\title{integrateItPack}
\description{
The integrateItPack package performs numeric integration using trapezoid or simpson rule
}
\examples{


}
\author{
Mauricio Vela: \email{mauriciovelabaron@wustl.edu}
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}, \code{\link{integrateIt}}, \code{\link{tolTest}}
}

