\name{integrateItpackage-package}
\alias{integrateItpackage-package}
\alias{integrateItpackage}
\docType{package}
\title{
\packageTitle{integrateItpackage}
}
\description{
\packageDescription{integrateItpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{integrateItpackage}
\packageIndices{integrateItpackage}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{integrateItpackage}

Maintainer: \packageMaintainer{integrateItpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
