% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simpson.R
\docType{class}
\name{Simpson-class}
\alias{Simpson-class}
\alias{Simpson-method}
\alias{initialize,Simpson-method}
\alias{plot,}
\alias{print,Simpson-method}
\title{A Simpson object}
\description{
Objects of class \code{Simpson} are created using \code{integrateIt} function.
}
\details{
An object of the class 'Simpson' has the following slots:
\itemize{
\item \code{x} Vector of x values
\item \code{y} Vector of y values
\item \code{a} Lower limit 
\item \code{b} Upper limit 
}
}
\author{
Mauricio Vela
}

