% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolTest.R
\name{tolTest}
\alias{tolTest}
\alias{tolTest,ANY-method}
\title{Numeric integration with degree of tolerance}
\usage{
tolTest(fun, a, b, start, tol, rule, correct, abserror)
}
\arguments{
\item{fun}{Function .}

\item{a}{Lower limit}

\item{b}{Upper limit}

\item{start}{initial value of n}

\item{tol}{level of tolerance}

\item{rule}{Simpson's Rule or the Trapezoidal Rule}

\item{correct}{A correct argument that provides the correct answer for the integral}
}
\value{
Returns final n, result for integration, and error
}
\description{
Function that increases the number of intervals until reaching a level of tolerance
}
\examples{

}
\author{
Mauricio Vela
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}, \code{\link{integrateIt}}
}

