% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{Trapezoid-method}
\alias{initialize,Trapezoid-method}
\alias{plot,}
\alias{print,Trapezoid-method}
\title{A Trapezoid object}
\description{
Objects of class \code{Trapezoid} are created using \code{integrateIt} function.
}
\details{
An object of the class 'Trapezoid' has the following slots:
\itemize{
\item \code{x} Vector of x values
\item \code{y} Vector of y values
\item \code{a} Lower limit 
\item \code{b} Upper limit 
}
}
\author{
Mauricio Vela
}

